fixed integration bugs
