+++ genrl/distributed/__init__.py
@@ -1,4 +1,4 @@
-from genrl.distributed.core import Master, DistributedTrainer, Node
+from genrl.distributed.core import Master, Node
 from genrl.distributed.parameter_server import ParameterServer, WeightHolder
 from genrl.distributed.experience_server import ExperienceServer
 from genrl.distributed.actor import ActorNode
