+++ genrl/trainers/__init__.py
@@ -3,3 +3,4 @@ from genrl.trainers.base import Trainer  # noqa
 from genrl.trainers.classical import ClassicalTrainer  # noqa
 from genrl.trainers.offpolicy import OffPolicyTrainer  # noqa
 from genrl.trainers.onpolicy import OnPolicyTrainer  # noqa
+from genrl.trainers.distributed import DistributedTrainer  # noqa
